+++ RAW_unity/Assets/_RAW/Scripts/Network/Inventory/NetworkInventoryEntry.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 52e31af614950471fa73000d4d020010
